Another pull

+++ database/migrations/2021_03_05_084637_create_carts_table.php
@@ -17,7 +17,7 @@ class CreateCartsTable extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('products_id');
-            $table->unsignedBigInteger('quantity')->default(1);
+            $table->unsignedTinyInteger('quantity')->default(1);
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users');
